[SWM-144] refactor: corrtect typo "reponse" -> "response"
